+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,6 +5,8 @@
 import FlutterMacOS
 import Foundation
 
+import raw_gnss
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  RawGnssPlugin.register(with: registry.registrar(forPlugin: "RawGnssPlugin"))
 }