+++ lib/src/skyplot.dart
@@ -2,6 +2,7 @@ import 'package:advance_math/advance_math.dart';
 import 'package:flutter/material.dart';
 
 import 'options.dart';
+import 'radar_options.dart';
 import 'sky_data.dart';
 import 'sky_object.dart';
 import 'sky_plot_painter.dart';
@@ -20,7 +21,7 @@ import 'sky_plot_painter.dart';
 ///   options: SkyPlotOptions(),
 /// )
 /// ```
-class SkyPlot extends StatelessWidget {
+class SkyPlot extends StatefulWidget {
   /// The list of sky data representing sky objects to be visualized.
   final List<SkyData> skyData;
 
@@ -29,6 +30,12 @@ class SkyPlot extends StatelessWidget {
 
   /// Options to customize the appearance and behavior of the sky plot.
   final SkyPlotOptions options;
+
+  /// Options to customize the appearance and behavior of the sky plot.
+  final RadarSweepOptions radarSweepOptions;
+
+  /// Whether to show radar animation
+  final bool showRadar;
 
   /// Creates a [SkyPlot] widget.
   ///
@@ -45,8 +52,51 @@ class SkyPlot extends StatelessWidget {
     super.key,
     required this.skyData,
     SkyPlotOptions? options,
+    RadarSweepOptions? radarSweepOptions,
     required this.categories,
-  }) : options = options ?? SkyPlotOptions();
+    this.showRadar = false,
+  })  : options = options ?? SkyPlotOptions(),
+        radarSweepOptions = radarSweepOptions ?? RadarSweepOptions();
+
+  @override
+  State<SkyPlot> createState() => _SkyPlotState();
+}
+
+class _SkyPlotState extends State<SkyPlot> with SingleTickerProviderStateMixin {
+  late AnimationController _controller;
+  late Animation<double> _animation;
+
+  @override
+  void initState() {
+    super.initState();
+    _controller = AnimationController(
+      duration: widget.radarSweepOptions.duration,
+      vsync: this,
+    )..repeat();
+
+    _animation = Tween<double>(begin: 0.0, end: 360.0).animate(_controller)
+      ..addListener(() {
+        setState(() {});
+      });
+  }
+
+  @override
+  void didUpdateWidget(SkyPlot oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    if (widget.showRadar != oldWidget.showRadar) {
+      if (widget.showRadar) {
+        _controller.repeat();
+      } else {
+        _controller.stop();
+      }
+    }
+  }
+
+  @override
+  void dispose() {
+    _controller.dispose();
+    super.dispose();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -54,27 +104,42 @@ class SkyPlot extends StatelessWidget {
       builder: (context, constraints) {
         var size = Size.square(constraints.biggest.shortestSide);
         var center = Offset(size.width / 2, size.height / 2);
-        var radius = size.width / 2;
+        // var radius = size.width / 2;
+
+        double maxRadius = size.width / 2;
+
+        // Reduce the radius if directional texts are shown to ensure they fit within the canvas
+        if (widget.options.showDirectionalText) {
+          maxRadius -= widget.options
+              .directionHorizontalTextOffset; // Adjust this value as needed
+        }
+
+        // Use this adjusted radius for drawing the elements
+        final radius = maxRadius - 5; // Additional padding for safety
 
         // Filter skyData based on the option
-        List<SkyData> filteredSkyData = options.showSkyObjectNotInFix
-            ? skyData
-            : skyData.where((data) => data.usedInFix ?? false).toList();
+        List<SkyData> filteredSkyData = widget.options.showSkyObjectNotInFix
+            ? widget.skyData
+            : widget.skyData.where((data) => data.usedInFix ?? false).toList();
 
         return Stack(
           children: [
             CustomPaint(
               painter: SkyPlotPainter(
-                skyData,
+                filteredSkyData,
                 context,
-                options,
+                widget.options,
+                _animation.value,
+                widget.showRadar,
+                widget.radarSweepOptions,
               ),
               size: size,
+              //
             ),
             ...filteredSkyData.map((skyObjectData) {
               double sizeOfFlag = (skyObjectData.usedInFix ?? false
-                      ? options.baseRadiusInFix
-                      : options.baseRadiusNotInFix) *
+                      ? widget.options.baseRadiusInFix
+                      : widget.options.baseRadiusNotInFix) *
                   2;
 
               // Convert azimuth and elevation to Cartesian coordinates
@@ -93,8 +158,8 @@ class SkyPlot extends StatelessWidget {
                 child: SkyObject(
                   skyData: skyObjectData,
                   size: sizeOfFlag,
-                  options: options,
-                  child: categories[skyObjectData.category]!,
+                  options: widget.options,
+                  child: widget.categories[skyObjectData.category]!,
                 ),
               );
             }),